requirements.txt
streamlit
openai
plotly
pandas
fpdf
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv